cloud_storage: rename manifest into base_manifest
