+++ src/v/archival/ntp_archiver_service.h
@@ -12,7 +12,6 @@
 #include "archival/archival_policy.h"
 #include "archival/probe.h"
 #include "archival/types.h"
-#include "cloud_storage/manifest.h"
 #include "cloud_storage/remote.h"
 #include "cloud_storage/types.h"
 #include "cluster/partition.h"
