+++ src/v/cloud_storage/CMakeLists.txt
@@ -3,7 +3,7 @@ v_cc_library(
   NAME cloud_storage
   SRCS
     cache_service.cc
-    manifest.cc
+    base_manifest.cc
     topic_manifest.cc
     partition_manifest.cc
     recursive_directory_walker.cc
