+++ src/v/cloud_storage/base_manifest.h
