+++ src/v/cloud_storage/tests/base_manifest_test.cc
@@ -8,7 +8,7 @@
  * https://github.com/vectorizedio/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_storage/manifest.h"
+#include "cloud_storage/base_manifest.h"
 #include "cloud_storage/types.h"
 #include "model/metadata.h"
 #include "seastarx.h"
