+++ src/v/cloud_storage/tests/offset_translation_layer_test.cc
@@ -8,7 +8,6 @@
  * https://github.com/vectorizedio/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_storage/manifest.h"
 #include "cloud_storage/offset_translation_layer.h"
 #include "cloud_storage/types.h"
 #include "seastarx.h"
